refactor: move components folder out of pages
